udpate notification

+++ back-end/src/handler/util/gcs.py
@@ -14,6 +14,9 @@ PACKAGE_LIMIT_GB = 10
 
 
 def move_file(src_bucket_name: str, src_filename: str, dest_bucket_name: str, dest_filename: str) -> None:
+    if get_env() == Env.DEV:
+        logger.debug(f'Skip gcs operations in env: DEV')
+        return
     storage_client = storage.Client()
 
     source_bucket = storage_client.bucket(src_bucket_name)
@@ -30,6 +33,9 @@ def move_file(src_bucket_name: str, src_filename: str, dest_bucket_name: str, de
 
 
 def _delete_file(bucket_name: str, filename: str) -> None:
+    if get_env() == Env.DEV:
+        logger.debug(f'Skip gcs operations in env: DEV')
+        return
     storage_client = storage.Client()
     bucket = storage_client.bucket(bucket_name)
     bucket.blob(filename).delete()
@@ -37,6 +43,9 @@ def _delete_file(bucket_name: str, filename: str) -> None:
 
 
 def _delete_folder(bucket_name: str, folder_path: str) -> None:
+    if get_env() == Env.DEV:
+        logger.debug(f'Skip gcs operations in env: DEV')
+        return
     storage_client = storage.Client()
     bucket = storage_client.get_bucket(bucket_name)
     blobs = bucket.list_blobs(prefix=folder_path)
